+++ Source/shared/Requests/Show.swift
@@ -36,15 +36,22 @@ public class TraktRequestShowProgress: TraktRequest {
         super.init(path: "/shows/\(showId)/progress/watched", params: params, oAuth: true)
     }
 
-    public func request(trakt: Trakt, completion: ([TraktSeason]?, NSError?) -> Void) -> Request? {
+    /**
+     Request show progress
+
+     - parameter trakt:      trakt client
+     - parameter completion: closure return the next episode to watch and all the seasons/episodes, NSError
+
+     - returns: Alamofire.Request
+     */
+    public func request(trakt: Trakt, completion: (TraktEpisode?, [TraktSeason]?, NSError?) -> Void) -> Request? {
         return trakt.request(self) { response in
             guard let data = response.result.value as? JSONHash,
-//                nextEpisode = data["next_episode"] as? JSONHash,
+                nextEpisode = TraktEpisode(data: data["next_episode"] as? JSONHash),
                 seasons = data["seasons"] as? [JSONHash] else {
-                return completion(nil, response.result.error)
+                return completion(nil, nil, response.result.error)
             }
-            // TraktEpisode(data: nextEpisode),
-            completion(seasons.flatMap {
+            completion(nextEpisode, seasons.flatMap {
                 TraktSeason(data: $0)
             }, nil)
         }